CRUD SAMPUN

+++ database/migrations/2025_02_28_151915_create_midtrans_logs_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('transaction_id');
             $table->string('order_id');
             $table->string('status_code', 10);
-            $table->decimal('gross_amount', 15, 2);
+            $table->integer('gross_amount');
             $table->string('payment_type', 50);
             $table->enum('transaction_status', ['pending', 'settlement', 'deny', 'expire', 'cancel', 'refund', 'partial_refund', 'chargeback']);
             $table->enum('fraud_status', ['accept', 'deny', 'challenge'])->default('accept');
